New changes to LA goggles/Binoculars behavior based on cvars, also brought back battery behavior which was cut from the final game

+++ codeJK2/game/NPC_reactions.cpp
@@ -974,7 +974,7 @@ void NPC_UseResponse( gentity_t *self, gentity_t *user, qboolean useWhenDone )
 NPC_Use
 -------------------------
 */
-extern void Add_Batteries( gentity_t *ent, int *count );
+extern void Add_Batteries( gentity_t *ent, float *count );
 
 void NPC_Use( gentity_t *self, gentity_t *other, gentity_t *activator ) 
 {
